feat: 폴더링 재구성, 코드 리팩토링 및 일부 기능 추가
